requirements.txt
pytorch-lightning
monai
einops
# timm # for BIT - Swin Transformer
# mmcv-full==1.6.0 # for ChangeFormer, BIT - ResNet
